Force redirects with conditions before plain redirects
