+++ lib/steps/setupRedirects.js
@@ -5,7 +5,7 @@ const {
   CUSTOM_REDIRECTS_PATH,
   NEXT_IMAGE_FUNCTION_NAME,
 } = require("../config");
-const getSortedRoutes = require("../helpers/getSortedRoutes");
+const getSortedRedirects = require("../helpers/getSortedRedirects");
 const getNetlifyRoutes = require("../helpers/getNetlifyRoutes");
 const isRootCatchAllRedirect = require("../helpers/isRootCatchAllRedirect");
 const isDynamicRoute = require("../helpers/isDynamicRoute");
@@ -49,13 +49,6 @@ const setupRedirects = (publishPath) => {
     target: `/.netlify/functions/${NEXT_IMAGE_FUNCTION_NAME}?url=:url&w=:width&q=:quality`,
   });
 
-  // Sort routes: More-specific routes precede less-specific routes (e.g., catch-all)
-  const getSortedRedirects = (redirects) => {
-    const sortedRoutes = getSortedRoutes(redirects.map(({ route }) => route));
-    return redirects.sort(
-      (a, b) => sortedRoutes.indexOf(a.route) - sortedRoutes.indexOf(b.route)
-    );
-  };
   const sortedStaticRedirects = getSortedRedirects(staticRedirects);
   const sortedDynamicRedirects = getSortedRedirects(dynamicRedirects);
